Changelog 10/12/2022:
Refactored large amounts of code relating to database and its table / UI derivatives. Ontop of bad variable names the code was calling the database an unecessary amount of times. The tables themselves did not reflect the structure agreed upon during planning.

UI was tweaked to not have overlapping elements, aswell as have extra input fields for relevant information. Whitelist now has a send frequency for example, and settings has a field for values.

Added table "Cache", for allowing caching of info recieved from client apps. Currently has no UI for viewing this data, possible future addition.

Added destrucitve versioning. As nothing stored in the database is mission critical enough to have to survive versioning it now deletes the contents when moving to a new version if it needs to. This is to prevent app crashing on major database alterations.